add modularity to kubernetes functions
