+++ internal/util/kubernetes/kubernetes_util.go
@@ -1,0 +1,32 @@
+package kubernetes
+
+import (
+	"regexp"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/util/validation"
+)
+
+// this file is designed to hold utility functions for Kubernetes that are not specific to
+// any type of resource
+
+// validManifestExtensions contains the supported extension for raw file.
+var validManifestExtensions = map[string]struct{}{"yaml": {}, "yml": {}, "json": {}}
+
+func IsValidKubernetesNamespace(name string) bool {
+	// All namespace names must be valid RFC 1123 DNS labels.
+	errs := validation.IsDNS1123Label(name)
+	reservedNamesRegex := regexp.MustCompile(`^(kubernetes-|kube-)`)
+	if len(errs) == 0 && !reservedNamesRegex.MatchString(name) {
+		return true
+	}
+	return false
+}
+
+func IsValidKubernetesManifestFile(fileName string) bool {
+	fileExt := strings.Split(fileName, ".")
+	if _, ok := validManifestExtensions[fileExt[len(fileExt)-1]]; ok {
+		return true
+	}
+	return false
+}
