+++ internal/util/kubernetes/structured_util_test.go
@@ -1,0 +1,37 @@
+package kubernetes
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+)
+
+func TestGetSecret(t *testing.T) {
+	scheme := runtime.NewScheme()
+	err := corev1.AddToScheme(scheme)
+	assert.NoError(t, err)
+	fakeClient := fake.NewClientBuilder().WithScheme(scheme).WithObjects(
+		&corev1.Secret{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test-secret",
+				Namespace: "test-namespace",
+			},
+			Data: map[string][]byte{
+				"key": []byte("value"),
+			},
+		},
+	).Build()
+
+	ctx := context.TODO()
+
+	secret, err := GetSecret(ctx, fakeClient, "test-namespace", "test-secret")
+
+	assert.NoError(t, err)
+	assert.NotNil(t, secret)
+	assert.Equal(t, "value", string(secret.Data["key"]))
+}
